% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITEVizTestData.R
\name{CITEVizTestData}
\alias{CITEVizTestData}
\title{Download a PBMC CITE-Seq data for CITEViz in 
Seurat/SingleCellExperiment format using ExperimentHub}
\usage{
CITEVizTestData(format = "Seurat")
}
\arguments{
\item{format}{Either "Seurat" or "SingleCellExperiment"}
}
\value{
A Seurat [default] or SingleCellExperiment object.
}
\description{
This function helps download a CITE-Seq dataset from 
Hao. et. al. 2021 for use with CITEViz. The dataset contains 
2500 cells and 228 antibodies.
}
\examples{
library(CITEVizTestData)
library(Seurat)
file_seurat <- CITEVizTestData(format = "Seurat")
file_sce <- CITEVizTestData(format = "SingleCellExperiment")


}
